feat: update Game Response to be more detailed
The changes made include updating the GameResponse_Full to have the developer,
publisher, gameImages, and discount responses included. This was done in an
effort to create a more complete response for the client, which can be used
to display a more detailed view of a game. This change also updates the
controller to include the relevant responses.

These updates aim to improve the information provided to the client and allow
for a more robust user experience.